Removed unnecessary library imports
