--- backend/example.env
@@ -1,5 +1,0 @@
-MONGO_HOST=your_mongo_host
-MONGO_PORT=your_mongo_port
-MONGO_AUTH_SOURCE=your_auth_source
-MONGO_USERNAME=your_username
-MONGO_PASSWORD=your_password
